# Improve performances when Prometheus metrics are enabled