--- a/pkg/metrics/prometheus.go
+++ b/pkg/metrics/prometheus.go
@@ -414,7 +414,8 @@ func (ps *prometheusState) Collect(ch chan<- stdprometheus.Metric) {
 }
 
 // DeletePartialMatch deletes all metrics where the variable labels contain all of those
-// passed in as labels. The order of the labels does not matter.
+// passed in as labels.
+// The order of the labels does not matter.
 // It returns the number of metrics deleted.
 func (ps *prometheusState) DeletePartialMatch(labels stdprometheus.Labels) int {
 	var count int