### What does this PR do?

Improves performances with metrics

Before:
```
Running 10s test @ http://traefik/bench
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.35ms    1.69ms  50.80ms   83.96%
    Req/Sec     5.83k   285.41     6.16k    97.00%
  Latency Distribution
     50%    4.15ms
     75%    4.97ms
     90%    6.03ms
     99%    7.67ms
  232189 requests in 10.01s, 22.59MB read
Requests/sec:  23187.77
Transfer/sec:      2.26MB
```

After:
```
Running 10s test @ http://traefik/bench
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.52ms    1.64ms  16.84ms   73.47%
    Req/Sec    10.49k     1.14k   30.77k    99.25%
  Latency Distribution
     50%    2.19ms
     75%    3.32ms
     90%    4.67ms
     99%    7.97ms
  418339 requests in 10.10s, 40.69MB read
Requests/sec:  41421.32
Transfer/sec:      4.03MB
```

### Motivation

Increase performances

Fixes #8118

### More

- [x] Added/updated tests
~- [ ] Added/updated documentation~

### Additional Notes

 Co-authored-by: Ludovic Fernandez <ldez@users.noreply.github.com>